Split shell protocol tests
